+++ assets/js/modules/adsense/components/dashboard/AdSenseDashboardWidget.js
@@ -29,7 +29,7 @@ import { __, _x, _n, sprintf } from '@wordpress/i18n';
 import Data from 'googlesitekit-data';
 import AdSenseIcon from '../../../../../svg/adsense.svg';
 import AdSensePerformanceWidget from './AdSensePerformanceWidget';
-import DashboardAdSenseTopPages from './DashboardAdSenseTopPages';
+import LegacyDashboardAdSenseTopPages from './LegacyDashboardAdSenseTopPages';
 import getNoDataComponent from '../../../../components/legacy-notifications/nodata';
 import getDataErrorComponent from '../../../../components/legacy-notifications/data-error';
 import ProgressBar from '../../../../components/ProgressBar';
@@ -189,7 +189,7 @@ export default function AdSenseDashboardWidget() {
 						</Cell>
 
 						<Cell className={ wrapperClass } size={ 12 }>
-							<DashboardAdSenseTopPages />
+							<LegacyDashboardAdSenseTopPages />
 						</Cell>
 
 						<Cell alignRight size={ 12 }>
