+++ assets/js/modules/adsense/components/dashboard/LegacyDashboardAdSenseTopPages.js
